renamed string and list scoring type modules
